idm/requirements.txt
grpcio
grpcio-tools
python-dotenv
redis
pyjwt
peewee
mysqlclient
fastapi
pydantic
uvicorn